+++ examples/use_rust/nrf52840/src/main.rs
@@ -65,7 +65,7 @@ async fn main(_spawner: Spawner) {
     // Initialize the storage and keymap
     let mut default_keymap = keymap::get_default_keymap();
     let storage_config = StorageConfig::default();
-    let mut behavior_config = BehaviorConfig::default();
+    let mut behavior_config = BehaviorConfig::<{ ROW }, { COL }>::default();
     let (keymap, mut storage) =
         initialize_keymap_and_storage(&mut default_keymap, flash, &storage_config, &mut behavior_config).await;